bug:33094 update codes to get rid of deprecation warnings

+++ trunk/mapreduce/src/java/com/marklogic/tree/CompressedTreeDecoder.java
@@ -663,7 +663,7 @@ public class CompressedTreeDecoder {
 
     private void decodeBinary(Decoder decoder, ExpandedTree rep, int nbytes) 
     throws IOException {
-        int nwords = (int) ((nbytes+3)/4);
+        int nwords = ((nbytes+3)/4);
         if (nwords <= 0) {
             LOG.error("nbytes=" + nbytes + ", nwords=" + nwords);
         }
